+++ lib/fog-octocloud.rb
@@ -1,2 +1,3 @@
 require 'fog'
+require 'fog/core_ext'
 require 'fog/octocloud'
