+++ lib/fog/octocloud.rb
@@ -9,23 +9,3 @@ module Fog
 
   end
 end
-
-# Monkey patch out provider lookup
-module Fog
-  module Compute
-    class << self
-      alias_method :pre_octocloud_new, :new
-
-      def new(attributes)
-        dup_attr = attributes.dup # prevent delete from having side effects
-        provider = dup_attr.delete(:provider).to_s.downcase.to_sym
-        if provider == :octocloud
-          require 'fog/octocloud/compute'
-          Fog::Compute::Octocloud.new(dup_attr)
-        else
-          pre_octocloud_new(attributes)
-        end
-      end
-    end
-  end
-end
